fixing errors and retrying deployment

+++ bin/entry_point.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-CONFIG_FILE=_config.yml 
+CONFIG_FILE=_config.yml
 
 /bin/bash -c "rm -f Gemfile.lock && exec jekyll serve --watch --port=8080 --host=0.0.0.0 --livereload --verbose --trace --force_polling"&
 
@@ -13,7 +13,11 @@ while true; do
     echo "Change detected to $CONFIG_FILE, restarting Jekyll"
 
     jekyll_pid=$(pgrep -f jekyll)
-    kill -KILL $jekyll_pid
+
+    # Check if jekyll_pid is not empty before killing
+    if [ -n "$jekyll_pid" ]; then
+      kill -KILL $jekyll_pid
+    fi
 
     /bin/bash -c "rm -f Gemfile.lock && exec jekyll serve --watch --port=8080 --host=0.0.0.0 --livereload --verbose --trace --force_polling"&
